backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==4.6.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.7.7
Pillow==10.1.0
requests==2.31.0
tweepy==4.14.0
facebook-sdk==3.1.0
httpx==0.25.2
psycopg2-binary==2.9.9
django-environ==0.11.2
celery[redis]==5.3.4
kombu==5.3.3
beautifulsoup4==4.12.2
feedparser==6.0.10
newspaper3k==0.2.8
boto3==1.28.85
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
whitenoise==6.6.0
gunicorn==21.2.0
django-cors-headers==4.3.1
